rename package `strackdriver` to `stackdriver`
