+++ modules/stackdriver-http-exporter/src/main/scala/io/janstenpickle/trace4cats/stackdriver/StackdriverHttpSpanCompleter.scala
@@ -1,4 +1,4 @@
-package io.janstenpickle.trace4cats.strackdriver
+package io.janstenpickle.trace4cats.stackdriver
 
 import cats.effect.{Blocker, Concurrent, ConcurrentEffect, Resource, Timer}
 import fs2.Chunk
@@ -7,8 +7,8 @@ import org.typelevel.log4cats.slf4j.Slf4jLogger
 import io.janstenpickle.trace4cats.`export`.{CompleterConfig, QueuedSpanCompleter}
 import io.janstenpickle.trace4cats.kernel.SpanCompleter
 import io.janstenpickle.trace4cats.model.TraceProcess
-import io.janstenpickle.trace4cats.strackdriver.oauth.TokenProvider
-import io.janstenpickle.trace4cats.strackdriver.project.ProjectIdProvider
+import io.janstenpickle.trace4cats.stackdriver.oauth.TokenProvider
+import io.janstenpickle.trace4cats.stackdriver.project.ProjectIdProvider
 import org.http4s.client.Client
 
 object StackdriverHttpSpanCompleter {
