+++ modules/stackdriver-http-exporter/src/main/scala/io/janstenpickle/trace4cats/stackdriver/model/SDAttributeValue.scala
@@ -1,4 +1,4 @@
-package io.janstenpickle.trace4cats.strackdriver.model
+package io.janstenpickle.trace4cats.stackdriver.model
 
 import io.circe.{Encoder, Json, JsonObject}
