+++ modules/stackdriver-http-exporter/src/main/scala/io/janstenpickle/trace4cats/stackdriver/oauth/package.scala
@@ -1,4 +1,4 @@
-package io.janstenpickle.trace4cats.strackdriver
+package io.janstenpickle.trace4cats.stackdriver
 
 import io.circe.generic.extras.Configuration
