+++ modules/stackdriver-http-exporter/src/main/scala/io/janstenpickle/trace4cats/stackdriver/project/InstanceMetadataProjectIdProvider.scala
@@ -1,4 +1,4 @@
-package io.janstenpickle.trace4cats.strackdriver.project
+package io.janstenpickle.trace4cats.stackdriver.project
 
 import cats.effect.Sync
 import org.http4s.Method.GET
